refactor: refactor decorator generator
