+++ libs/alpha/ui/antd/src/components/grid/Grid.input.ts
@@ -1,4 +1,5 @@
 import { Layout } from 'react-grid-layout'
+import { Grid } from '@codelab/tools/generators/json-schema'
 
 // export type ReactGridResponsiveLayout = Pick<
 //   ResponsiveProps,
@@ -11,4 +12,20 @@ export type ReactGridResponsiveLayout = {
   breakpoints: Record<LayoutSize, number>
   cols: Record<LayoutSize, number>
   layouts: Record<LayoutSize, Layout>
+}
+
+export class DemoGridFormInput {
+  @Grid({ order: 3, span: 24 })
+  email = ''
+
+  @Grid({ order: 2, span: 24 })
+  password = ''
+
+  @Grid({ order: 0, span: 12 })
+  firstname = ''
+
+  @Grid({ order: 1, span: 12 })
+  lastname = ''
+
+  notGroupedField = ''
 }
