+++ libs/modules/app/src/core/application/useCases/DemoGridForm/DemoGridFormInput.ts
@@ -1,13 +1,13 @@
-import { grid } from '@codelab/tools/generators/json-schema'
+import { Grid } from '@codelab/tools/generators/json-schema'
 
 export class DemoGridFormInput {
-  @grid({ order: 1, span: 12 })
+  @Grid({ order: 1, span: 12 })
   email1 = ''
 
-  @grid({ order: 0, span: 16 })
+  @Grid({ order: 0, span: 16 })
   password0 = ''
 
-  @grid({ order: 2, span: 8 })
+  @Grid({ order: 2, span: 8 })
   name2 = ''
 
   notGroupedField = ''
