+++ libs/tools/generators/json-schema/src/decorators/index.ts
@@ -1,1 +1,1 @@
-export * from './gridDecorator'
+export * from './decorator-grid'
