+++ workspace.json
@@ -1471,7 +1471,7 @@
           "builder": "@nrwl/workspace:run-commands",
           "options": {
             "commands": [
-              "npx ts-node --project libs/tools/generators/json-schema/tsconfig.lib.json -r tsconfig-paths/register libs/tools/generators/json-schema/src/generate.ts"
+              "npx ts-node --project libs/tools/generators/json-schema/tsconfig.lib.json -r tsconfig-paths/register libs/tools/generators/json-schema/src/main.ts"
             ],
             "parallel": false
           }
